+++ src/PhpSpreadsheet/Reader/Slk.php
@@ -268,6 +268,7 @@ class Slk extends BaseReader
         if ($sharedFormula === true && $sharedRow >= 0 && $sharedColumn >= 0) {
             $thisCoordinate = Coordinate::stringFromColumnIndex((int) $column) . $row;
             $sharedCoordinate = Coordinate::stringFromColumnIndex($sharedColumn) . $sharedRow;
+            /** @var string */
             $formula = $spreadsheet->getActiveSheet()->getCell($sharedCoordinate)->getValue();
             $spreadsheet->getActiveSheet()->getCell($thisCoordinate)->setValue($formula);
             $referenceHelper = ReferenceHelper::getInstance();
@@ -281,6 +282,7 @@ class Slk extends BaseReader
             return;
         }
         $columnLetter = Coordinate::stringFromColumnIndex((int) $column);
+        /** @var string */
         $cellData = Calculation::unwrapResult($cellData);
 
         // Set cell value
@@ -404,7 +406,7 @@ class Slk extends BaseReader
                 $endCol = Coordinate::stringFromColumnIndex((int) $endCol);
                 $spreadsheet->getActiveSheet()->getColumnDimension($startCol)->setWidth((float) $columnWidth);
                 do {
-                    $spreadsheet->getActiveSheet()->getColumnDimension(++$startCol)->setWidth((float) $columnWidth);
+                    $spreadsheet->getActiveSheet()->getColumnDimension((string) ++$startCol)->setWidth((float) $columnWidth);
                 } while ($startCol !== $endCol);
             }
         }