+++ src/PhpSpreadsheet/Shared/Escher/DgContainer/SpgrContainer.php
@@ -32,6 +32,8 @@ class SpgrContainer
 
     /**
      * Add a child. This will be either spgrContainer or spContainer.
+     *
+     * @param SpgrContainer|SpgrContainer\SpContainer $child child to be added
      */
     public function addChild(mixed $child): void
     {