+++ src/PhpSpreadsheet/Style/NumberFormat.php
@@ -2,6 +2,8 @@
 
 namespace PhpOffice\PhpSpreadsheet\Style;
 
+use PhpOffice\PhpSpreadsheet\RichText\RichText;
+
 class NumberFormat extends Supervisor
 {
     // Pre-defined formats
@@ -453,7 +455,7 @@ class NumberFormat extends Supervisor
     /**
      * Convert a value in a pre-defined format to a PHP string.
      *
-     * @param mixed $value Value to format
+     * @param null|bool|float|int|RichText|string $value Value to format
      * @param string $format Format code: see = self::FORMAT_* for predefined values;
      *                          or can be any valid MS Excel custom format string
      * @param ?array $callBack Callback function for additional formatting of string